requirements.txt
streamlit
python-hcl2
pyyaml
fpdf